requirements.txt
discord.py==1.7.3
asyncio==3.4.3
requests==2.26.0
pymongo==3.12.0
dnspython==2.1.0
gTTS==2.2.3
PyNaCl==1.4.0
wikipedia==1.4.0
youtube-dl==2021.6.6
discord-pretty-help==1.3.3
svglib==1.1.0
akinator.py==5.0.0
jishaku==2.3.0
